requirements.txt
# Core Dependencies
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Deep Learning Frameworks
torch>=2.0.0  # Will use MPS backend on Apple Silicon
torchvision>=0.15.0
torchaudio>=2.0.0

# Reinforcement Learning
stable-baselines3>=2.0.0
gymnasium>=0.28.0
shimmy>=1.0.0  # For gym compatibility

# Time Series & Financial Analysis
ta>=0.10.2  # Technical indicators
ta-lib>=0.4.25  # Advanced technical analysis
pandas-ta>=0.3.14b0
yfinance>=0.2.18
ccxt>=4.0.0  # Cryptocurrency exchange library
python-binance>=1.0.17
hurst>=0.0.5

# AutoML & Optimization
flaml>=2.0.0
ray[tune]>=2.3.0
optuna>=3.2.0
hyperopt>=0.2.7

# Classic ML Models
lightgbm>=4.0.0
xgboost>=1.7.0
catboost>=1.2.2

# Transformers & Advanced Models
transformers>=4.30.0
einops>=0.6.1  # For transformer operations
timm>=0.9.0  # Pretrained models

# Data Processing & Augmentation
numba>=0.57.0  # JIT compilation
bottleneck>=1.3.7  # Fast array operations
pyarrow>=12.0.0  # Efficient data storage
fastparquet>=0.8.3

# Visualization & Monitoring
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
dash>=2.10.0  # For interactive dashboards
rich>=13.3.0  # Beautiful terminal output

# MLOps & Experiment Tracking
mlflow>=2.4.0
wandb>=0.15.0
tensorboard>=2.13.0

# Database & Storage
sqlalchemy>=2.0.0
redis>=4.5.0
pymongo>=4.3.0  # For MongoDB support

# API & Web
fastapi>=0.100.0
uvicorn>=0.22.0
websockets>=11.0
aiohttp>=3.8.0
requests>=2.31.0

# Scheduling
apscheduler>=3.10.4

# Testing & Quality
pytest>=7.3.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.3.0
flake8>=6.0.0
mypy>=1.3.0

# Testing dependencies
pytest==7.4.3
pytest-cov==4.1.0
pytest-benchmark==4.0.0
psutil==5.9.6
memory-profiler==0.61.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
typer>=0.9.0  # CLI creation
tqdm>=4.65.0
joblib>=1.3.0
dill>=0.3.6
cloudpickle>=2.2.0

# Apple Silicon Specific (install separately if on M-series Mac)
# mlx>=0.5.0
# mlx-lm>=0.0.6

# Async Support
asyncio>=3.4.3
aiofiles>=23.1.0
aiodns>=3.0.0

# Cryptography & Security
cryptography>=41.0.0
python-jose[cryptography]>=3.3.0
passlib>=1.7.4

# Data Validation
marshmallow>=3.19.0
cerberus>=1.3.4

# Time Handling
python-dateutil>=2.8.2
pytz>=2023.3
tzlocal>=5.0

# Configuration Management
pyyaml>=6.0
toml>=0.10.2
configparser>=6.0.0

# Logging & Monitoring
loguru>=0.7.0
sentry-sdk>=1.25.0  # Error tracking

# Performance Monitoring
py-spy>=0.3.14
memory-profiler>=0.60.0
line-profiler>=4.0.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.2.0
mkdocs>=1.4.0
mkdocs-material>=9.1.0

# Development Tools
ipython>=8.14.0
jupyter>=1.0.0
jupyterlab>=4.0.0
ipywidgets>=8.0.0

# Optional: Advanced Features
# tensorflow>=2.13.0  # For comparison/ensemble
# jax>=0.4.13  # For advanced optimization
# pennylane>=0.31.0  # Quantum-inspired algorithms
